Fix Logger - externalize in helpers
