+++ src/app.ts
@@ -1,10 +1,11 @@
 import express from 'express';
-import { createLogger, format, transports } from 'winston';
 import connectDatadog from 'connect-datadog';
 import cors from 'cors';
 import helmet from 'helmet';
 import bodyParser from 'body-parser';
+import logger from './helpers/logger';
 import dotenv from 'dotenv';
+dotenv.config();
 
 import candidatureRouter from './routes/candidature';
 import evenementRouter from './routes/evenement';
@@ -12,13 +13,6 @@ import qcmRouter from './routes/qcm';
 import entrepriseRouter from './routes/entreprise';
 import offreRouter from './routes/offre';
 
-dotenv.config();
-
-
-const dd_options = {
-  'response_code': true,
-  'tags': [`app:api-admitech-${process.env.NODE_ENV}`]
-};
 
 const app = express();
 const port = 3000;
@@ -29,37 +23,13 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(helmet());
 
-
-// Logger creation
-const logger = createLogger({
-  level: 'info',
-  format: format.combine(
-    format.timestamp({
-      format: 'YYYY-MM-DD HH:mm:ss'
-    }),
-    format.errors({ stack: true }),
-    format.splat(),
-    format.json()
-  ),
-  defaultMeta: { service: `api-admitech-${process.env.NODE_ENV}` },
-  transports: [
-    new transports.File({ filename: 'logs/test.log' })
-  ]
-});
-// If we're not in production then **ALSO** log to the `console`
-// with the colorized simple format.
-//
-if (process.env.NODE_ENV !== 'production') {
-  logger.add(new transports.Console({
-    format: format.combine(
-      format.colorize(),
-      format.simple()
-    )
-  }));
-}
+const dd_options = {
+  'response_code': true,
+  'path': true,
+  'tags': [`app:api-admitech-${process.env.NODE_ENV}`]
+};
 
 app.use(connectDatadog(dd_options));
-
 
 // Primary routes
 app.use('/candidature', candidatureRouter);
